requirements.txt
flask
crewai[google-genai]
python-dotenv
werkzeug

